+++ kaye/gen_prompt/prompt_blueprint.py
@@ -6,17 +6,75 @@ import re
 from datetime import datetime
 
 import importlib.metadata
-from anytree import RenderTree, PreOrderIter
+from anytree import RenderTree, PreOrderIter, Node
+from anytree.render import ContStyle
 
-from .prompt_corpus_node import ROOT_NODE_NAME
+from .. import kamilog, PROGRAM_NAME
+from .prompt_corpus_node import HEADING_PREFIX
 
 __all__ = ("PromptBlueprint",)
 
+logger = kamilog.getLogger(PROGRAM_NAME)
 
-# types of line prefix used for __repr__() generation
-CHECKED_BOX_PREFIX = "[x] "
-UNCHECKED_BOX_PREFIX = "[ ] "
-NO_CHECKBOX_PREFIX = "    "
+
+class _PreviewTreeNode(Node):
+    """
+    helper class usedVjj
+    in ``.generate_preview_tree()`` of ``class PromptBlueprint``
+
+    create a tree structured same as prompt corpus tree
+    for the utility of preview tree generation
+
+
+    :param blueprint: blueprint object which create this preview tree
+    :type blueprint: PromptBlueprint
+    :param concurrent_corpus_node: concurrent node in the prompt corpus tree
+    :type concurrent_corpus_node: PromptCorpusNode
+    :param parent:
+    :type parent: _PreviewTreeNode
+    """
+
+    def is_enabled(self):
+        """
+        :return: whether this node is enabled in the blueprint;
+        :rtype: bool
+        """
+        # todo opmz by hash names_path
+        names_path = self.concurrent_corpus_node.names_path
+        return any(
+            (names_path == node.names_path) for node in self.blueprint.enabled
+        )
+
+    def prune_trivial_branches(self):
+        """
+        prune all branches that contains no enabled nodes
+
+        :return: whether this node has any enabled descendants
+        :rtype: bool
+        """
+        if self.is_leaf:
+            if self.is_enabled():
+                return True  # keep this enabled leaf node
+        else:
+            non_trivial = [
+                child.prune_trivial_branches() for child in self.children
+            ]
+            if self.is_enabled() or any(non_trivial):
+                return True  # self is marked/children has marked nodes
+
+        # remove self from tree, ready for garbage collection
+        self.parent = None
+        return False
+
+    def __init__(self, blueprint, concurrent_corpus_node, parent):
+        super().__init__(concurrent_corpus_node.name, parent)
+
+        self.blueprint = blueprint
+        self.concurrent_corpus_node = concurrent_corpus_node
+
+        # create children nodes
+        for child_concurrent_node in concurrent_corpus_node.children:
+            _PreviewTreeNode(blueprint, child_concurrent_node, self)
 
 
 class PromptBlueprint:
@@ -27,122 +85,96 @@ class PromptBlueprint:
     A ``PromptBlueprint`` mirrors the hierarchical structure of the prompt
     corpus, but each node can be explicitly **enabled** or **disabled**.
 
-    Use ``__repr__()`` to generate a visual representation of the **tree**.
+    Use ``.generate_preview_tree()`` (or ``__repr__()``) to generate
+    a visual representation of the **tree**
+    showing enabled node with ``[x]`` and disabled node with ``[ ]``
 
-    Use ``__str__()`` to render a **concrete prompt** composed of nodes.
-    supports 2 operational modes:
-
-    - **detached mode**: Any node is included in the output if it is enabled,
-      regardless of its parent or ancestor status.
-
-    - else: A node appears in the output
-      if **both** it and all its ancestor nodes are enabled
+    Use ``.generate_prompt()`` (or ``__str__()``) to render
+    a **concrete prompt** composed of all enabled nodes
 
 
-    :param prompt_corpus: *prompt corpus* tree root node
+    :param prompt_corpus: *prompt corpus* tree **root** node
             which this prompt blueprint attached to
     :type prompt_corpus: PromptCorpusNode
-    :param blueprint_name: display name given to the prompt
-    :type blueprint_name: str, optional
-    :param prompt_blueprint_text: prompt blueprint text to set nodes.
-            It must be formatted identical to output of ``__repr__()``
-            (with tree structure and checkboxes.)
-            if ``None``:
-            the created ``PromptBlueprint`` has *all* nodes **disabled**
-    :type prompt_blueprint_text: str, optional
-    :param detached_mode: defaults to False
-    :type detached_mode: bool, optional
+    :param blueprint_text: prompt blueprint text to set nodes,
+            must in the same format of output of ``__repr__()``
+            (with tree structure and checkboxes;)
+            if ``None``: create an **empty** prompt blueprint,
+            i.e. all nodes disabled
+    :type blueprint_text: str
+    :param display_name: display name given to the prompt,
+            defaults to ""
+    :type display_name: str, optional
     """
 
     @classmethod
     def create_full_prompt_blueprint(
-        cls, prompt_corpus, blueprint_name="full"
+        cls, prompt_corpus, blueprint_display_name="full"
     ):
         """
         :param prompt_corpus: *prompt corpus* tree root node
                 which this prompt blueprint attached to
         :type prompt_corpus: PromptCorpusNode
-        :param prompt_blueprint_name: display name given to the prompt;
+        :param blueprint_display_name: display name given to the prompt;
                 defaults to "full"
-        :type prompt_blueprint_name: str, optional
+        :type blueprint_display_name: str, optional
         :return: an instance of ``PromptBlueprint`` attached to the given
                 ``prompt_corpus``, and with **all nodes enabled**
         :rtype: PromptBlueprint
         """
-        blueprint = cls(prompt_corpus, prompt_blueprint_name=blueprint_name)
+        blueprint = cls(
+            prompt_corpus, blueprint_display_name=blueprint_display_name
+        )
         # set all nodes
         for node in PreOrderIter(prompt_corpus):
-            if node is prompt_corpus:  # skip root node
+            if node.parent is None:  # skip root node
                 continue
 
-            blueprint.enabled_nodes_names.append(node.name)
+            blueprint.enabled.append(node)
 
         return blueprint
 
-    @property
-    def is_detached_mode(self):
-        """
-        :return: whether the PromptBlueprint is in **detached mode**
-        :rtype: bool
-        """
-        return ROOT_NODE_NAME not in self.enabled_nodes_names
-
-    def set_unset_detached_mode(self, detached_mode):
-        """
-        Set or unset the **detached mode** for the PromptBlueprint.
-
-        :param detached_mode: whether to set or unset the detached mode.
-        :type detached_mode: bool
-        """
-        if detached_mode:  # set detached mode
-            # ensure "○" is absent in ``.enabled_nodes_names``
-            if ROOT_NODE_NAME in self.enabled_nodes_names:
-                self.enabled_nodes_names.remove(ROOT_NODE_NAME)
-
-        else:  # unset detached mode
-            # ensure "○" is present in ``.enabled_nodes_names``
-            if ROOT_NODE_NAME not in self.enabled_nodes_names:
-                self.enabled_nodes_names.append(ROOT_NODE_NAME)
-
     def generate_preview_tree(
         self,
+        *,
+        show_full_tree=False,
         preview_line_count=3,
         preview_line_width=64,
-        *,
-        hide_comment=False
+        hide_comment=False,
     ):
         """
-        Return a visual representation of the **tree**, showing:
+        generate a visual representation of the **tree**, showing:
 
         - tree structure
         - node name (i.e. section heading)
-        - node enabled/disabled status indicated by checkboxes
+        - node enabled/disabled status, prefixed with:
+
+          - ``[x]`` for enabled node
+          - ``[ ]`` for disabled node
+
         - node content preview
 
-        Root node is named ``○``.
-        If it is prefixed with unchecked checkbox ``[ ]``,
-        the instance operate in *detached mode*.
 
-
+        :param show_full_tree: _description_, defaults to False
+        :type show_full_tree: bool, optional
         :param preview_line_count: set maximum line count of
                 *content preview* part for each entry,
-                (excluding section heading line)
+                (excluding section heading line;)
                 defaults to 3
-        :type preview_line_count: int
+        :type preview_line_count: int, optional
         :param preview_line_width: set maximum column width of
                 *content preview* part for each entry;
-                defaults to 64.
-        :type preview_line_width: int
-        :param hide_comment: Disable placing comment part after last line;
-                Defaults to False
+                defaults to 64
+        :type preview_line_count: int, optional
+        :param hide_comment: disable comment part after last line;
+                defaults to False
         :type hide_comment: bool, optional
         :return: v.s.
         :rtype: str
-
         :example:
         >>> tree = PromptBlueprint(...)
         >>> tree.generate_preview_tree()
-        [x] ○
+            ○
         [x] └── Project Title
         [ ]     ├── Description
                 │   A brief overview of the project, its purpose, and goals.
@@ -160,7 +192,7 @@ class PromptBlueprint:
                     This project is licensed under the MIT License.
         (blueprint:conversation; Kaye v1.2.3)
         >>> tree.generate_preview_tree(preview_line_count=0, hide_comment=True)
-        [x] ○
+            ○
         [x] └── Project Title
         [ ]     ├── Description
         [ ]     ├── Installation
@@ -169,42 +201,43 @@ class PromptBlueprint:
         [x]     └── License
         """
 
+        preview_tree = _PreviewTreeNode(self, self.prompt_corpus, None)
+
+        if not show_full_tree:
+            preview_tree.prune_trivial_branches()
+
         opt_lines = []
+        for pre, fill, node in RenderTree(preview_tree, style=ContStyle()):
+            if node.parent is None:
+                root_line = "    {}".format(node.name)
+                opt_lines.append(root_line)
+                continue
 
-        for pre, fill, node in RenderTree(self.prompt_corpus):
-            node_name = node.name
-            # decide either have [x] or [ ] before node lines
-            checkbox_prefix = (
-                CHECKED_BOX_PREFIX
-                if node_name in self.enabled_nodes_names
-                else UNCHECKED_BOX_PREFIX
-            )
+            # create node / heading line
+            checkbox = "[x] " if node.is_enabled() else "[ ] "
+            heading_line = checkbox + pre + node.name
+            opt_lines.append(heading_line)
 
-            opt_lines.append(checkbox_prefix + pre + node_name)
-
-            # lines for the content of node
+            # generate content preview
             opt_lines.extend(
-                node.generate_preview_tree_content_part(
-                    NO_CHECKBOX_PREFIX + fill,
-                    preview_line_count,
-                    preview_line_width,
+                node.concurrent_corpus_node.generate_preview_tree_content_part(
+                    "    " + fill, preview_line_count, preview_line_width
                 )
             )
 
+        # append comment line
         if not hide_comment:
-            comment_line = "(" + self._generate_prompt_comment_content() + ")"
+            comment_line = "({})".format(self._generate_comment_content())
             opt_lines.append(comment_line)
 
         return "\n".join(opt_lines)
 
     def generate_prompt(self, *, hide_comment=False):
         """
-        :param hide_comment: disable placing comment part after last line;
+        :param hide_comment: disable comment part after last line;
                 defaults to False
         :type hide_comment: bool, optional
-        :return: **concrete prompt** composed of nodes heading and content,
-                depending on *detached mode* and each nodes' enabling status.
-                Q.v. ``PromptBlueprint``
+        :return: **concrete prompt** composed of nodes heading and content
         :rtype: str
         :example:
         >>> tree = PromptBlueprint(...)
@@ -216,97 +249,133 @@ class PromptBlueprint:
         ## Conclusion
         Summarizing the findings and implications.
         """
-        lines = (
-            self._generate_str_recursively_detached_mode(self.prompt_corpus)
-            if self.is_detached_mode
-            else self._generate_str_recursively(self.prompt_corpus)
-        )
+        # generate lines from root node
+        lines = self._generate_prompt_recursively(self.prompt_corpus)
 
         # create comment part
         if not hide_comment:
-            comment_line = (
-                "<!-- " + self._generate_prompt_comment_content() + " -->"
-            )
+            comment_line = "<!-- " + self._generate_comment_content() + " -->"
             lines.append(comment_line)
 
         return "\n".join(lines)
 
+    HEADING_LINE_PATTERN = r"\[([x ])\] (.*)[└├]── (.+)"
+
     def __init__(
         self,
         prompt_corpus,
-        prompt_blueprint_text=None,
-        prompt_blueprint_name=None,
-        detached_mode=False,
+        blueprint_text=None,
+        *,
+        display_name="",
     ):
-        self.blueprint_name = prompt_blueprint_name
+        self.display_name = display_name
         self.prompt_corpus = prompt_corpus
-        self.enabled_nodes_names = []  # all nodes currently enabled
 
-        if prompt_blueprint_text:
-            self._init_populate_enabled_nodes_names(prompt_blueprint_text)
-        else:
-            # enable tree root node means non-detached mode
-            self.set_unset_detached_mode(detached_mode)
+        # list of all enabled nodes
+        self.enabled = []  # default as empty blueprint
 
-    def _init_populate_enabled_nodes_names(self, prompt_blueprint_text):
-        lines = prompt_blueprint_text.split("\n")
+        if blueprint_text:
+            self._init_populate_enabled_by_blueprint_text(blueprint_text)
 
-        # parse detached mode
-        self.set_unset_detached_mode(
-            lines[0] != CHECKED_BOX_PREFIX + ROOT_NODE_NAME
-        )
+    def _init_populate_enabled_by_blueprint_text(self, blueprint_text):
+        """
+        helper method used in ``__init__()``
 
-        # find all node names in the tree
-        node_names = [node.name for node in self.prompt_corpus.descendants]
+        populate ``self.enabled`` by parsing the init param ``blueprint_text``
+        """
+        lines = blueprint_text.split("\n")
+
+        path2node = {
+            node.names_path: node for node in self.prompt_corpus.descendants
+        }
 
         # extract all enabled headings
-        pattern = r"{}.+── (.+)".format(re.escape(CHECKED_BOX_PREFIX))
+        previous_level = -1
+        previous_path = []
         for line in lines:
-            match = re.fullmatch(pattern, line)
-            if match:  # find a line start w/ checked box
-                found_heading = match.group(1)
-                if found_heading in node_names:
-                    # ensure the found heading in present in the tree
-                    self.enabled_nodes_names.append(found_heading)
+            match = re.fullmatch(self.HEADING_LINE_PATTERN, line)
+            # find node / heading as a line
+            if match:
+                is_checked = match.group(1) == "x"
+                level = len(match.group(2)) // 4
+                heading = match.group(3)
 
-    def _generate_str_recursively(self, node):
-        # stop recursive if this node is not enabled
-        if node.name not in self.enabled_nodes_names:
-            return []
+                # dynamically decide the names_path
+                if level > previous_level:
 
-        lines = []
-        if node.parent is not None:  # skip root node
-            lines.extend(node.generate_heading_and_content_lines())
+                    if level - previous_level > 1:
+                        logger.error(
+                            "detect bad blueprint tree format at:\n%s", line
+                        )
+                        continue
 
-        # children
-        for child_node in node.children:
-            lines.extend(self._generate_str_recursively(child_node))
+                    path = previous_path + [""]
 
-        return lines
+                elif level == previous_level:
+                    path = previous_path
 
-    def _generate_str_recursively_detached_mode(self, node):
-        lines = []
+                else:
+                    path = previous_path[: level + 1]
 
-        if node.name in self.enabled_nodes_names and node.parent is not None:
-            lines.extend(node.generate_heading_and_content_lines())
+                path[level] = heading
 
-        # children
-        for child_node in node.children:
-            lines.extend(
-                self._generate_str_recursively_detached_mode(child_node)
-            )
+                path_tuple = tuple(path)
+                if path_tuple not in path2node:
+                    logger.warning(
+                        "not part of the provided prompt corpus, skipped"
+                        " during blueprint parsing:\n%s",
+                        line,
+                    )
+                    continue
 
-        return lines
+                if is_checked:
+                    node = path2node[path_tuple]
+                    self.enabled.append(node)
 
-    def _generate_prompt_comment_content(self):
+                previous_level, previous_path = level, path
+
+    def _generate_prompt_recursively(self, node):
         """
-        :return: prompt comment (used in __repr__() and __str__()) containing
-                blueprint name and Kaye version
-        :rtype: str
+        helper method used in ``.generate_preview()``
 
+        generate recursively prompt lines from ``node``
+        """
+        lines = []
+
+        try:
+            idx = self.enabled.index(node)
+        except ValueError:
+            idx = -1
+        if idx >= 0 and node.parent is not None:
+            level = node.depth
+
+            # add empty lines before headings
+            if idx > 0:
+                lines.append("")
+
+            # add heading line
+            heading_line = HEADING_PREFIX * level + " " + node.name
+            lines.append(heading_line)
+            # add content lines
+            lines.extend(node.content)
+
+        # children
+        for child_node in node.children:
+            lines.extend(self._generate_prompt_recursively(child_node))
+
+        return lines
+
+    def _generate_comment_content(self):
+        """
+        helper method used in ``.generate_preview_tree()`` and
+        ``.generate_prompt()``
+
+
+        :return: prompt comment containing blueprint name and Kaye version
+        :rtype: str
         :example:
         >>> print(tree._generate_prompt_comment_content())
-        'blueprint:conversation; Kaye v1.2.3'
+        'blueprint: chat; Kaye v1.2.3'
         """
         kaye_version = importlib.metadata.version("kaye")
 
@@ -314,14 +383,13 @@ class PromptBlueprint:
         if "a" in kaye_version:
             kaye_version += datetime.now().strftime(".0%Y%m%d%H%M%S")
 
-        return "{}Kaye v{}".format(
-            (
-                "blueprint:{}; ".format(self.blueprint_name)
-                if self.blueprint_name
-                else ""
-            ),
-            kaye_version,
+        name_part = (
+            "blueprint: {}; ".format(self.display_name)
+            if self.display_name
+            else ""
         )
+
+        return "{}Kaye v{}".format(name_part, kaye_version)
 
     def __repr__(self):
         return self.generate_preview_tree()